==> finalized/3_Full_IDAC_Input.ipynb <==
import numpy as np
import pandas as pd
import pickle
import oenotebook as oenb
from openeye.oechem import *
from openmoltools.openeye import *
from openeye.oeiupac import *

#from extraction import main
from test_extract import main

### Part 1: Checking SD Tags of Orion Output Data

#Use output .oeb file and the following lines to create OEMol object with SD Tags
input_file = 'output/results/IDAC_solvation_success.oeb'
its = oechem.oemolistream(input_file)
molecule = oechem.OEMol()
oechem.OEReadMolecule(its, molecule)

tag_name = 'name'
#Check if a Certain SD Tag exists and its value
oechem.OEGetSDData(molecule, tag_name)

def DumpSDData(mol):
    print("SD data of {}:".format(mol.GetTitle()))
    #Loop over SD Data
    for dp in oechem.OEGetSDDataPairs(mol):
        print(dp.GetTag())

##Use imported function to convert .oeb input into a .csv
# main('output/results/IDAC_solvation_success.oeb', 'output/results/IDAC_solvation_success.csv')
# main('output/results/IDAC_solvation_fail.oeb', 'output/results/IDAC_solvation_fail.csv')
# main('output/results/IDAC_ssolv_success.oeb', 'output/results/IDAC_ssolv_success.csv')
# main('output/results/IDAC_ssolv_fail.oeb', 'output/results/IDAC_ssolv_fail.csv')

def counting(table):
    i = 0
    for key in table:
        i += len(table[key])
    return i

#Read in output files from Orion
solv_success = pd.read_csv('output/results/IDAC_solvation_success.csv')
solv_fail = pd.read_csv('output/results/IDAC_solvation_fail.csv')

#Remove leading whitespace in column names
solv_success.columns = [x.lstrip() for x in list(solv_success.columns)]
solv_fail.columns = [x.lstrip() for x in list(solv_fail.columns)]

ssolv_success = pd.read_csv('output/results/IDAC_ssolv_success.csv')
ssolv_fail = pd.read_csv('output/results/IDAC_ssolv_fail.csv')

#Remove leading whitespace in column names
ssolv_success.columns = [x.lstrip() for x in list(ssolv_success.columns)]
ssolv_fail.columns = [x.lstrip() for x in list(ssolv_fail.columns)]

#Combined list of solvation and self-solvation calculation outputs
combined_solv = solv_fail.append(solv_success)
combined_ssolv = ssolv_fail.append(ssolv_success)

#Import pickled dictionary to assess which entries are missing
temp_dict = pickle.load(open('temp_dict.p', 'rb'))
ref_table = pickle.load(open('input/final_table.p', 'rb'))
solv_table = pickle.load(open('input/final_solv.pkl', 'rb'))
ssolv_table = pickle.load(open('input/final_ssolv.pkl', 'rb'))

#Create dictionaries with all compounds and the temperatures they were calculated at
solv_dict = {}
for name in combined_solv.Molecule.unique():
    subset = combined_solv[combined_solv.Molecule == name]
    solv_dict[name] = [x for x in subset['Temperature(K)']]

ssolv_dict = {}
for name in combined_ssolv.Molecule.unique():
    subset = combined_ssolv[combined_ssolv.Molecule == name]
    ssolv_dict[name] = [x for x in subset['Temperature(K)']]

#Convert compound names to smiles to compare between dictionaries 
smiles = {}
for name in temp_dict:
    mol_from_name = OEMol()
    OEParseIUPACName(mol_from_name, name)
    smiles[name] = OECreateCanSmiString(mol_from_name)

smiles_solv = {}
for name in solv_dict:
    mol_from_name = OEMol()
    OEParseIUPACName(mol_from_name, name)
    smiles_solv[name] = OECreateCanSmiString(mol_from_name)
    
smiles_ssolv = {}
for name in ssolv_dict:
    mol_from_name = OEMol()
    OEParseIUPACName(mol_from_name, name)
    smiles_ssolv[name] = OECreateCanSmiString(mol_from_name)

#Create dictionaries with keys as SMILES strings for comparison
total_dict = {}
for name in temp_dict:
    total_dict[smiles[name]] = temp_dict[name]
    
new_solv_dict = {}
for name in solv_dict:
    new_solv_dict[smiles_solv[name]] = solv_dict[name]

new_ssolv_dict = {}
for name in ssolv_dict:
    new_ssolv_dict[smiles_ssolv[name]] = ssolv_dict[name]

print('Temp_dict:', counting(total_dict))
print('Solvation_dict:', counting(new_solv_dict))
print('Self-Solvation_dict:', counting(new_ssolv_dict))

### Searching for Dropped Values within Keys

dropped_solv = {}
for key in new_solv_dict:
    if sorted(total_dict[key]) != sorted(new_solv_dict[key]):
        dropped_solv[key] = list(set(total_dict[key]) ^ set(new_solv_dict[key]))
        
dropped_ssolv = {}        
for key in new_ssolv_dict:
    if sorted(total_dict[key]) != sorted(new_ssolv_dict[key]):
        dropped_ssolv[key] = list(set(total_dict[key]) ^ set(new_ssolv_dict[key]))

print(counting(dropped_solv))
print(counting(dropped_ssolv))

### Searching for Dropped Keys

original_smiles = set([key for key in total_dict])
solv_smiles = set([key for key in new_solv_dict])
ssolv_smiles = set([key for key in new_ssolv_dict])

solv_difference = list(original_smiles ^ solv_smiles)
ssolv_difference = list(original_smiles ^ ssolv_smiles)

bob = []
for smile in solv_difference:
    dropped_solv[smile] = total_dict[smile]

for smile in ssolv_difference:
    dropped_ssolv[smile] = total_dict[smile]

print(counting(dropped_solv))
print(counting(dropped_ssolv))

### Printing Dropped Calculations

solv_rip = pd.DataFrame()
for key in dropped_solv:
    subset = solv_table[solv_table['solute_SMILES'] == key]
    for temp in dropped_solv[key]:
        solv_rip = solv_rip.append(subset[subset.temperature == temp])

ssolv_rip = pd.DataFrame()
for key in dropped_ssolv:
    subset = ssolv_table[ssolv_table.solvents == key]
    for temp in dropped_ssolv[key]:
        ssolv_rip = ssolv_rip.append(subset[subset.temperature == temp])

pd.DataFrame.to_pickle(solv_rip, 'error/dropped_solv.pkl')
pd.DataFrame.to_pickle(ssolv_rip, 'error/dropped_ssolv.pkl')

### Adjusting Size of Solvation and Self-Solvation Sets

solv_success.head()
len(solv_success)

ssolv_success.head()
len(ssolv_success)

#Generate lists of all compounds and the calculated temperatures for solv and self-solv
solv_yes = []
for name in solv_success.Molecule.unique():
    subset = solv_success[solv_success.Molecule == name]
    for temp in subset['Temperature(K)']:
        solv_yes.append(str(name) + ';' + str(temp))
        
ssolv_yes = []
for name in ssolv_success.Molecule.unique():
    subset = ssolv_success[ssolv_success.Molecule == name]
    for temp in subset['Temperature(K)']:
        ssolv_yes.append(str(name) + ';' + str(temp))

A = set(solv_yes)
B = set(ssolv_yes)

diff = A^B
print(len(diff))
print(len(A-diff))
print(len(B-diff))

### Remove the Calculations without Both Solvation and Self-Solvation

adjusted_solv = list(A-diff)
adjusted_ssolv = list(B-diff)

new_solv = pd.DataFrame()
for entry in adjusted_solv:
    subset = solv_success[solv_success.Molecule == entry.split(';')[0]]
    new_solv = new_solv.append(subset[subset['Temperature(K)'] == float(entry.split(';')[1])])

print(len(new_solv))
new_solv.head()

new_ssolv = pd.DataFrame()
for entry in adjusted_ssolv:
    subset = ssolv_success[ssolv_success.Molecule == entry.split(';')[0]]
    new_ssolv = new_ssolv.append(subset[subset['Temperature(K)'] == float(entry.split(';')[1])])
    
new_ssolv['Solvents(smiles)'] = [x.lstrip() for x in new_ssolv['Solvents(smiles)']]

print(len(new_ssolv))
new_ssolv.head()

ref_table.head()

##Re-insert experimental errors from IDAC paper
error_exp = []
for smile,temp in zip(new_ssolv['Solvents(smiles)'], new_ssolv['Temperature(K)']):
    subset = ref_table[ref_table['SMILES'] == smile]
    subset = subset[subset['Temp'] == temp]
    error_exp.append(float(subset['Uncertain'].item()))

new_ssolv['Uncertain_exp'] = error_exp
uncertain = new_ssolv['Uncertain_exp']
uncertain.unique()

new_ssolv[new_ssolv['Uncertain_exp'] == 30.]

ratios = []
len(new_ssolv[new_ssolv['Uncertain_exp'] == 0.])
for index, compound in enumerate(new_ssolv['Molecule']):
    subset = new_ssolv[new_ssolv['Molecule'] == compound]
    for i in range(len(subset)):
        if list(subset['Uncertain_exp'])[i] == 0:
            ratios.append('0' + ';' + str(compound))
        else:
            ratios.append(str(list(subset['Uncertain_exp'])[i]/list(subset['IDAC_expt'])[i]) + ';' + str(compound))

### Part 2: Calculating IDAC from Orion Output Data

Using the following formula: $$\gamma_{i}^{\infty} = exp\bigg(\frac{\Delta G_{i}^{solv} - \Delta G_{i}^{self-solv}}{k_{B}T} \bigg) \cdot \frac{\rho_{molar}^{solvent}}{\rho_{molar}^{pure-solute}} $$

def calc_idac(temperature, G_solvation, G_self_solvation, top_density, bottom_density):
    '''
    top_density: final density in solvation
    bottom_density: final density in self-solvation
    '''
    IDAC_calc = []
    k = 0.0019872041 #units in kcal/mol*K
    for i in range(len(temperature)):
        value = np.exp((G_solvation[i] - G_self_solvation[i])/(k*temperature[i])) * (top_density[i]/bottom_density[i])
        IDAC_calc.append('{:.2f}'.format(value))
    return IDAC_calc

def exp_error(temperature, IDAC_exp, error):
    error_exp = []
    k = 0.0019872041
    for i in range(len(temperature)):
        value = ((k*temperature[i])/IDAC_exp[i])**2 * error[i]**2
        error_exp.append('{:.5f}'.format(value))
    return error_exp

def calc_error(temperature, solv_error, ssolv_error, top_density, bottom_density, top_density_error, bottom_density_error):
    error_calc = []
    k = 0.0019872041
    for i in range(len(temperature)):
        value = solv_error[i]**2 + ssolv_error[i]**2 + ((k*temperature[i])/top_density[i])**2 * top_density_error[i]**2 + ((k*temperature[i])/bottom_density[i])**2 * bottom_density_error[i]**2
        error_calc.append('{:.2f}'.format(value))
    return error_calc

def calculate_IDAC(solvation, self_solvation):
    table = pd.DataFrame(columns = ['Name', 'Temp', 'IDAC_calc', 'IDAC_expt'])
    table['Name'] = solvation['Molecule']
    table['Temp'] = solvation['Temperature(K)']
    table['IDAC_calc'] = calc_idac([x for x in table['Temp']],
                                   [x for x in solvation['FE(kcal/mol)']],
                                   [x for x in self_solvation['FE(kcal/mol)']], 
                                   [x for x in solvation['Density_Final(g/ml)']],
                                   [x for x in self_solvation['Density_Final(g/ml)']])
    table['IDAC_expt'] = [str(x) for x in solvation['IDAC_expt']]
    table['kT*ln(IDAC_calc)_Error'] = calc_error([x for x in table['Temp']],
                                                 [x for x in solvation['DFE(kcal/mol)']],
                                                 [x for x in self_solvation['DFE(kcal/mol)']],
                                                 [x for x in solvation['Density_Final(g/ml)']],
                                                 [x for x in self_solvation['Density_Final(g/ml)']],
                                                 [x for x in solvation['DDensity_Final(g/ml)']],
                                                 [x for x in self_solvation['DDensity_Final(g/ml)']])
    
    table['kT*ln(IDAC_exp)_Error'] = exp_error([x for x in self_solvation['Temperature(K)']],
                                               [x for x in self_solvation['IDAC_expt']],
                                               [x for x in self_solvation['Uncertain_exp']])
    return table

new_solv.columns = [x.lstrip() for x in list(new_solv.columns)]
new_ssolv.columns = [x.lstrip() for x in list(new_ssolv.columns)]

final = calculate_IDAC(new_solv, new_ssolv)

final['kT*ln(IDAC_calc)_Error'] = [float(x) for x in final['kT*ln(IDAC_calc)_Error']]
final['kT*ln(IDAC_exp)_Error'] = [float(x) for x in final['kT*ln(IDAC_exp)_Error']]

final[final['IDAC_calc'] == '0.00'] 

#Drop the entries with calculated IDAC values of 0
final = final[final['IDAC_calc'] != '0.00']
final = final.reset_index(drop = True)

final.sort_values(by = ['Name', 'Temp']).head()

pd.DataFrame.to_pickle(final, 'output/IDAC_FINAL.pkl')

### old stuff

results = oenb.read_file_to_dataframe("output/results_final.oeb.gz")

smile = 'CCCCOC(=O)C'
sub_group = results[results['Solute SMILES'] == smile]

sub_group[sub_group['Solvent name'] == 'water']

calc_idac = np.exp((2.913815)/(0.0019872041*273.35))

col = ['Name', 'Temperature (K)', '$\Delta G_{sol}$ (kcal/mol)', '$\Delta G_{selfsolv}$ (kcal/mol)', 'IDAC_calc', 'IDAC_exp']
compare = pd.DataFrame(columns = col)
compare.append(pd.DataFrame([['Chris', 298.15, -3.293, -6.239, 161.2, 814.00], ['Guilherme', 273.35, -3.929, -6.931, 213.6, 613]], columns = col))
